Homogenize MediaSpatialApp sample with CustomComponentsSample
Summary:
As part of an overhaul I'm doing of the Spatial SDK Getting Started docs, I'm rewriting our LookAt tutorial to use both the MediaSpatialAppTemplate and CustomComponentsStarter apps.

Both repos have different file structures and different files, making it impossible to include them both in one document. I've opened this diff to copy the required files from the CustomComponentsStarter to the MediaSpatialAppTemplate so that integration is possible.

Reviewed By: zsp1987

Differential Revision: D68975417

fbshipit-source-id: 905f50efe4cce34ee2eb3d68621c38d32ec141e1