Update MediaAppTemplate repo for 0.5.5 release
Summary:
The 0.5.5 release changes component attributes files so they use XML instead of Kotlin. This diff updates the LookAt component in https://github.com/fbsamples/MediaSpatialAppTemplate to use XML and removes references to it from ids.xml, since as of 0.5.5 you don't need to register components in that file anymore.

I commented out LookAt.xml because it's originally from another one of our samples, and it causes errors if left uncommented. I copied it to this repo as part of my LookAt tutorial overhaul (D66474462), and the Gradle files aren't configured yet. The user configures them during the tutorial, which makes the error go away.

NOTE: This is for the 0.5.5 release (2/13/25) and should not be shipped until those changes are confirmed to be live.

Reviewed By: mwthot

Differential Revision: D69536507

fbshipit-source-id: 88bf0124c93803726d41a2c2062eb25f33cabd24